+++ plugins/quick-brick-google-ima-apple/apple/.jazzy.json
