+++ plugins/quick-brick-google-ima-apple/apple/ZappGoogleInteractiveMediaAds.podspec.json
