+++ plugins/quick-brick-google-ima-apple/apple/tvos/GoogleInteractiveMediaAds.framework/VERSION
