+++ plugins/quick-brick-google-ima-apple/apple/universal/GoogleInteractiveMediaAdsAdapter/GoogleInteractiveMediaAdsAdapter+IMAAdsManagerDelegate.swift
