+++ plugins/quick-brick-google-ima-apple/apple/universal/GoogleUrlTagData/MidrollTagData.swift
