+++ plugins/quick-brick-google-ima-apple/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@applicaster/quick-brick-google-ima-apple",
-  "version": "0.7.0",
+  "version": "0.10.4",
   "description": "",
   "main": "src/index.js",
   "scripts": {
@@ -8,21 +8,23 @@
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/applicaster/AppleApplicasterFrameworks"
+    "url": "git+https://github.com/applicaster/Zapp-Frameworks"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "bugs": {
-    "url": "https://github.com/applicaster/AppleApplicasterFrameworks"
+    "url": "https://github.com/applicaster/Zapp-Frameworks"
   },
-  "homepage": "https://github.com/applicaster/AppleApplicasterFrameworks",
+  "homepage": "https://github.com/applicaster/Zapp-Frameworks",
   "applicaster": {
     "isPlugin": true,
     "supportedPlatforms": [
       "ios_for_quickbrick",
-      "ios"
+      "ios",
+      "tvos_for_quickbrick",
+      "tvos"
     ],
-    "zappOwnerAccountId": "5e12fac7666333000ec0f8ab"
+    "zappOwnerAccountId": "5e12fe6948ca8800083c4116"
   }
 }
